# Add inline and inbounds annotations to unsafe_dot