@martinholters 2025-07-09
Any idea how much difference the `@inline` and the `@inbounds` each make on their own?