@jacobga1998 2025-07-09
It is mostly the inline. Rerunning the benchmark:

Without the @inbounds:
julia> @benchmark filt!(output, interp_fil, data)
BenchmarkTools.Trial: 190 samples with 1 evaluation per sample.
 Range (min … max):  21.867 ms … 44.836 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     24.969 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.243 ms ±  3.886 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▇ █▇▄▄▄ ▄▃▁
  ▅▅▆█▆█████▇████▇▅▁▃▃▁▄▆▁▁▄▃▆▃▅▅▅▅▄▄▃▄▁▅▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▃▁▁▃▃ ▃
  21.9 ms         Histogram: frequency by time        39.7 ms <

 Memory estimate: 16 bytes, allocs estimate: 1.

Without the @inline 

julia> @benchmark filt!(output, interp_fil, data)
BenchmarkTools.Trial: 124 samples with 1 evaluation per sample.
 Range (min … max):  33.064 ms … 54.623 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     39.826 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   40.553 ms ±  4.594 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

           ▁  ▅   █  ▂▄  ▄
  ▃▁▅█▁▅▅▆██▆▅█▆█▅█▅███▆▆██▃▆▃▁▃▃▁▆▃▃▅▁▃▁▅▅▆▃▁▅▃▁▃▁▃▁▅▁▁▁▁▁▁▅ ▃
  33.1 ms         Histogram: frequency by time        53.6 ms <

 Memory estimate: 16 bytes, allocs estimate: 1.