@martinholters 2025-07-09
Thanks. In my opinion, overriding the compiler heuristics needs good justification, and those benchmarks surely qualify. The additional `@inbounds` seem about as (un)safe as the existing ones, so seems fair to add them, too.