@martinholters 2025-07-09
> But if this helps only interpolation, maybe we can use call-site inlining for just that?

Good point. Can you try that, @jacobga1998?