@jacobga1998 2025-07-10
Agreed, that looks nicer. Added the suggestions.